removed none code files
